// lancewells_github_io/src/ItemData/Enums/WeaponProperties.ts
export enum WeaponProperties {
    Ammunition = "Ammunition",
    Finesse = "Finesse",
    Heavy = "Heavy",
    Improvised = "Improvised",
    Light = "Light",
    Loading = "Loading",
    Reach = "Reach",
    Silver = "Silver",
    Special = "Special",
    Thrown = "Thrown",
    TwoHanded = "Two-Handed",
    Versatile = "Versatile"
}